pyproject.toml
[build-system]
requires = ['setuptools', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = "mn-gsend"
requires-python = ">=3.7"

# author = "Markiyan Kushnr"
# author-email = "markiyan.kushnir@gmail.com"
description = "Mn GSend"

dependencies = [
    # "flask-restful",
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth",
]

dynamic = [
    "version",
]

[tool.setuptools]
packages = [
    "mngsend",
]

[tool.setuptools.dynamic]
version = {attr = "mngsend.version.__version__"}
